feat: Verify upstream image
If the upstream image is tampered with we should automatically fail the
image build and prevent any futher security breaches.

This is not foolproof as the implementation grabs the public key from a
git repository instead of locally, but it will ensure that the key from
the git repository did sign the image.

Fixes: #25
Co-authored-by: XLion <xlion@xlion.tw>
Signed-off-by: Freya Gustavsson <freya@venefilyn.se>